Revert "Origin/feature/it21365614 refine"

+++ src/main/java/com/example/onlineticketingsystem/repo/UserRepo.java
@@ -3,8 +3,5 @@ package com.example.onlineticketingsystem.repo;
 import com.example.onlineticketingsystem.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface UserRepo extends JpaRepository<User, Integer> {
-
+public interface UserRepo extends JpaRepository<User,Long> {
 }
